Extract SafeCmd stdin lifecycle tests into their own module (#117)
Address the CodeScene "Lines of Code in a Single File" finding on
`test_safe_cmd_run.py` with a cohesive test-module extraction.

- Add `cuprum/unittests/test_safe_cmd_stdin.py` holding the stdin writer
  lifecycle regression group moved verbatim (assertions, commands, payload
  size, timeout values, and async/sync coverage unchanged):
  `test_stdin_input_with_timeout_escalation`,
  `test_direct_timeout_with_blocked_stdin_writer_does_not_hang`, and
  `test_stdin_input_cancellation_cleans_up_task`. The module carries only the
  imports and local helpers (`_execute_async`/`_execute_sync`, a local
  `python_builder` fixture) those tests need; `collections.abc` is imported
  under `TYPE_CHECKING` since it is annotation-only here.
- Remove those tests from `test_safe_cmd_run.py`, retaining
  `test_streamed_run_cancellation_cleans_up_task` (stream-consumer cleanup via
  captured execution) and the shared runtime helpers/imports still used there.
- Refresh the maturin wheel-build snapshot for the new module.

No production subprocess code changed and no CodeScene suppression added;
timeout, blocked-drain, cancellation, and execution-strategy coverage are
preserved. `test_safe_cmd_run.py` drops from 635 to 565 non-blank,
non-comment lines.
